% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{run_task1}
\alias{run_task1}
\title{Run Simulation for Task 1}
\usage{
run_task1()
}
\value{
A data.frame with the mean and variance of misclassification errors.
}
\description{
Runs the simulation experiment for Task 1 (condition: x1 < x2, nodesize: 25).
}
