% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{run_task2}
\alias{run_task2}
\title{Run Simulation for Task 2}
\usage{
run_task2()
}
\value{
A data.frame with the mean and variance of misclassification errors.
}
\description{
Runs the simulation experiment for Task 2 (condition: x1 < 0.5, nodesize: 25).
}
