% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{run_task3}
\alias{run_task3}
\title{Run Simulation for Task 3}
\usage{
run_task3()
}
\value{
A data.frame with the mean and variance of misclassification errors.
}
\description{
Runs the simulation experiment for Task 3
(condition: (x1 < 0.5 & x2 < 0.5) | (x1 > 0.5 & x2 > 0.5), nodesize: 12).
}
